+++ CHANGELOG.rst
@@ -2,7 +2,19 @@
 Changelog
 =========
 
+0.2.1
+------------------
+- Added minimal toe clearance metric
+- Added stride length metric
+- Added MoS metric
+- Fixed bug for get events when no events entry is present in c3d file
+- Fixed bug where phase dependent metrics were added to their respective feature in the results
+- Fixed bug reversed progression vector
+- Fixed bug where the Gaitalytics logo did not show up in the documentation
+
 0.2.0
 ------------------
 Breaking changes:
 The whole library was rewritten. File formats from older version will not be compatible
+
+